pyproject.toml
[project]
name = "test-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi[all]>=0.118.0",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.43",
    "strawberry-graphql[cli,fastapi]>=0.283.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.3",
]

[tool.pytest.ini_options]
asyncio_default_test_loop_scope = 'session'
asyncio_default_fixture_loop_scope = 'session'
